[WIP] Update database schema and work on importer
Updated the schema to be aligned with the continous verse
ID schema. The importer will now update the database
to that schema and load the t_[ABV].json file into memory.

Next steps are to load the file into the database semi-efficently.
And then load the other cross references and such.