Update verse identification to allow for multi-verse reference.
Multi-verse references are `data-usfm` strings of the form:
  1CH.1.1+1CH.1.2

In these cases we will split the reference by '+', collect all verse
data in the FIRST verse of the range, and update the maximum number of
verses by the LAST verse of the range.

Also allow for verses that do not have a typical `fullText` parallel.