Better use of stores and state.
Switch from `update_state()` to using derived stores for `vers_abbr` and
`verse_ref`.

Use `verse_input` to bind to the input value, initialized by the
`verse_ref` store value.

Make use of the new `primary_version` store to ensure that any change
will be maintained when we leave and return to the page.