Tighten up fonts, update colors, put more info in popovers.
Wrap each verse in a <div class='verse'/> in preparation for presenting
multiple versions.

Use flexbox gap to separate items, both horizontally and vertically
within a verse vs using padding within each item.

Change the word container to a simple div with centered content
(no flexbox needed since all items are blocks).

Move the type-of-speech and Strongs reference into the Popover for the
english, removing the need to present it in the primary display.

Make the color of the verse label yellow to match VersionYvers.
Make the english text white, the wlc text blue, and the transliteration
grey.