[enhancement] Better management of selected state
Initialze the selected state in `reset_selecting()` using the current
`verse` data.

Change `afterUpdate()` to use the `selected_store` instead of the
current `verse` data when updating the set of selected verses.
This resolves a bug where navigation to a specific set of verse would
remove the ability to de-select verses.