Update load to indicate whether the data is new.
Always perform the get since, if we do not there will be issues on
either the server or client side.

BUT, update the content to include the target version and api ref along
with an indication of whether this is new/changed content.

This `changed` indicator will be used client-side to determine whether
`is_loading` should be set.